[cherry-pick][XPU] Refine the unit tests (#4457) test=develop
