+++ lite/tests/api/CMakeLists.txt
@@ -8,18 +8,17 @@ function(lite_cc_test_with_model_and_data TARGET)
     set(multiValueArgs "")
     cmake_parse_arguments(args "${options}" "${oneValueArgs}" "${multiValueArgs}" ${ARGN})
 
-    set(ARGS "")
     if(DEFINED args_MODEL)
-        set(ARGS "${ARGS} --model_dir=${LITE_MODEL_DIR}/${args_MODEL}")
+        set(ARGS ${ARGS} --model_dir=${LITE_MODEL_DIR}/${args_MODEL})
     endif()
     if(DEFINED args_DATA)
-        set(ARGS "${ARGS} --data_dir=${LITE_MODEL_DIR}/${args_DATA}")
+        set(ARGS ${ARGS} --data_dir=${LITE_MODEL_DIR}/${args_DATA})
     endif()
     if(DEFINED args_CONFIG)
-        set(ARGS "${ARGS} --config_dir=${LITE_MODEL_DIR}/${args_CONFIG}")
+        set(ARGS ${ARGS} --config_dir=${LITE_MODEL_DIR}/${args_CONFIG})
     endif()
     if(DEFINED args_ARGS)
-        set(ARGS "${ARGS} ${args_ARGS}")
+        set(ARGS ${ARGS} ${args_ARGS})
     endif()
     lite_cc_test(${TARGET} SRCS ${TARGET}.cc
         DEPS ${lite_model_test_DEPS} paddle_api_full
@@ -45,7 +44,7 @@ function(lite_cc_test_with_model_and_data TARGET)
 endfunction()
 
 if(LITE_WITH_ARM)
-    lite_cc_test_with_model_and_data(test_transformer_with_mask_fp32_arm MODEL transformer_with_mask_fp32 ARGS)
+    lite_cc_test_with_model_and_data(test_transformer_with_mask_fp32_arm MODEL transformer_with_mask_fp32)
 endif()
 
 if(LITE_WITH_NPU)
@@ -55,11 +54,11 @@ if(LITE_WITH_NPU)
 endif()
 
 if(LITE_WITH_XPU AND NOT LITE_WITH_XTCL)
-    lite_cc_test_with_model_and_data(test_resnet50_fp32_xpu MODEL resnet50 DATA ILSVRC2012_small)
-    lite_cc_test_with_model_and_data(test_googlenet_fp32_xpu MODEL GoogLeNet DATA ILSVRC2012_small)
-    lite_cc_test_with_model_and_data(test_vgg19_fp32_xpu MODEL VGG19 DATA ILSVRC2012_small)
-    lite_cc_test_with_model_and_data(test_ernie_fp32_xpu MODEL ernie DATA bert_data)
-    lite_cc_test_with_model_and_data(test_bert_fp32_xpu MODEL bert DATA bert_data)
+    lite_cc_test_with_model_and_data(test_resnet50_fp32_baidu_xpu MODEL resnet50 DATA ILSVRC2012_small)
+    lite_cc_test_with_model_and_data(test_googlenet_fp32_baidu_xpu MODEL GoogLeNet DATA ILSVRC2012_small)
+    lite_cc_test_with_model_and_data(test_vgg19_fp32_baidu_xpu MODEL VGG19 DATA ILSVRC2012_small)
+    lite_cc_test_with_model_and_data(test_ernie_fp32_baidu_xpu MODEL ernie DATA bert_data)
+    lite_cc_test_with_model_and_data(test_bert_fp32_baidu_xpu MODEL bert DATA bert_data)
 endif()
 
 if(LITE_WITH_RKNPU)
