+++ lite/tests/api/test_fpr_fp32_baidu_xpu.cc
@@ -26,7 +26,7 @@
 namespace paddle {
 namespace lite {
 
-TEST(ResnetCbam, test_resnet_cbam_fp32_xpu) {
+TEST(ResnetCbam, test_resnet_cbam_fp32_baidu_xpu) {
   lite_api::CxxConfig config;
   // config.set_model_dir(FLAGS_model_dir);
   config.set_model_file(FLAGS_model_dir + "/__model__");
