+++ lite/tests/api/test_resnet50_fp32_baidu_xpu.cc
@@ -32,7 +32,7 @@ DEFINE_int32(channel, 3, "image channel");
 namespace paddle {
 namespace lite {
 
-TEST(Resnet50, test_resnet50_fp32_xpu) {
+TEST(Resnet50, test_resnet50_fp32_baidu_xpu) {
   lite_api::CxxConfig config;
   config.set_model_dir(FLAGS_model_dir);
   config.set_valid_places({lite_api::Place{TARGET(kXPU), PRECISION(kFloat)},
